refactor: 테스트 코드 패키지 구조 변경
- data, ui 레이어에 맞게 테스트 코드의 패키지 구조를 변경했습니다.
- 이에 따라 관련 테스트 클래스의 import 경로를 수정했습니다.
