+++ app/src/test/java/woowacourse/shopping/ui/cart/CartActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.cart
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.ViewModelProvider
@@ -8,11 +8,9 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
 import org.robolectric.RobolectricTestRunner
-import woowacourse.shopping.ui.MainActivity
-import woowacourse.shopping.ui.MainViewModel
 
 @RunWith(RobolectricTestRunner::class)
-class MainActivityTest {
+class CartActivityTest {
     @get:Rule
     var instantTaskExecutorRule = InstantTaskExecutorRule()
 
@@ -21,7 +19,7 @@ class MainActivityTest {
         // given
         val activity =
             Robolectric
-                .buildActivity(MainActivity::class.java)
+                .buildActivity(CartActivity::class.java)
                 .create()
                 .get()
 
@@ -34,10 +32,10 @@ class MainActivityTest {
         // given
         val activity =
             Robolectric
-                .buildActivity(MainActivity::class.java)
+                .buildActivity(CartActivity::class.java)
                 .create()
                 .get()
-        val viewModel = ViewModelProvider(activity)[MainViewModel::class.java]
+        val viewModel = ViewModelProvider(activity)[CartViewModel::class.java]
 
         // then
         assertThat(viewModel).isNotNull()
